# docs/api-specification.yaml
...
openapi: 3.0.3
info:
  title: RocketVault API
  description: |
    A self-hosted, open-source alternative to Azure Key Vault: secrets
    management, cryptographic key operations, X.509 certificate lifecycle
    management, and multi-vault RBAC.
  version: 4.0.0
  contact:
    name: RocketVault
    url: https://github.com/Snehal1112/rocketvault
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.rocketvault.local
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/vaults:
    get:
      summary: List vaults
      description: |
        Returns every vault. Requires the admin role or a global `vaults/manage`
        allow policy — a vault-scoped grant is not sufficient.
      operationId: listVaults
      tags:
        - Vaults
      parameters:
        - name: include_deleted
          in: query
          required: false
          description: Include soft-deleted vaults when set to `true`
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Vaults retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListVaultsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create a vault
      description: |
        Creates a new vault. Requires the admin role or a global `vaults/manage`
        allow policy — a vault-scoped grant on another vault is not sufficient.
      operationId: createVault
      tags:
        - Vaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaultRequest"
      responses:
        "201":
          description: Vault created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/vaults/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Name of the vault
        schema:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$"
          example: "production"
    get:
      summary: Get a vault
      description: |
        Returns a vault by name. Authorized against the named vault itself, not
        the default vault.
      operationId: getVault
      tags:
        - Vaults
      responses:
        "200":
          description: Vault retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Update a vault
      description: |
        Updates a vault's enabled flag, purge protection, retention period, or
        tags. Nil fields are left unchanged; an empty tag map clears all tags.
      operationId: updateVault
      tags:
        - Vaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultRequest"
      responses:
        "200":
          description: Vault updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Soft-delete a vault
      description: |
        Soft-deletes a vault and cascades to its secrets, keys, and
        certificates. The default vault cannot be deleted.
      operationId: deleteVault
      tags:
        - Vaults
      responses:
        "204":
          description: Vault soft-deleted successfully
        "400":
          description: The default vault is protected from deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/vaults/{name}/webhook:
    parameters:
      - name: name
        in: path
        required: true
        description: Name of the vault
        schema:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$"
          example: "production"
    put:
      summary: Create or update a vault's webhook config
      description: |
        Creates or updates the vault's webhook notification target.
        Authorized against the named vault itself (`vaults/manage`), the same
        as the other `/vaults/{name}` management routes.

        The server always mints the signing secret; there is no
        client-supplied secret field. `rotate_secret: true` replaces an
        existing secret. `enabled` is omittable to leave the current value
        (or default to `true` on create) rather than silently disabling the
        webhook by omission.

        The response carries `signing_secret` in plaintext only on the call
        that mints one — the first create, or any update with
        `rotate_secret: true`. It is shown exactly once: the secret is stored
        encrypted, and no endpoint (including the GET below) ever returns it
        again after this response.
      operationId: upsertVaultWebhook
      tags:
        - Vaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVaultWebhookRequest"
      responses:
        "200":
          description: |
            Webhook config saved. The `signing_secret` field is present only
            when this call minted a secret (create, or update with
            `rotate_secret: true`); an update that did not rotate returns the
            config without it.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/VaultWebhookConfigCreatedResponse"
                  - $ref: "#/components/schemas/VaultWebhookConfigResponse"
              examples:
                created_or_rotated:
                  summary: Create or rotate — secret shown once
                  value:
                    url: "https://hooks.example.com/rocketvault"
                    enabled: true
                    created_at: "2026-08-19T12:00:00Z"
                    updated_at: "2026-08-19T12:00:00Z"
                    signing_secret: "b64-encoded-secret-shown-only-now"
                updated_without_rotate:
                  summary: Update without rotating — no secret field
                  value:
                    url: "https://hooks.example.com/rocketvault"
                    enabled: true
                    created_at: "2026-08-19T12:00:00Z"
                    updated_at: "2026-08-19T12:05:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: Get a vault's webhook config
      description: |
        Returns the vault's webhook config. The response never includes the
        signing secret in any form, encrypted or plain — the secret is shown
        exactly once, from the PUT call that minted it.
      operationId: getVaultWebhook
      tags:
        - Vaults
      responses:
        "200":
          description: Webhook config retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultWebhookConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The vault does not exist, or has no webhook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a vault's webhook config
      description: Removes the vault's webhook config. Deleting when none exists succeeds.
      operationId: deleteVaultWebhook
      tags:
        - Vaults
      responses:
        "204":
          description: Webhook config deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/vaults/{vault_name}/purge:
    parameters:
      - $ref: "#/components/parameters/VaultName"
    delete:
      summary: Purge a vault
      description: |
        Permanently removes a soft-deleted vault and everything it held. Gated
        purely by the vault-scoped role-assignment check, so even a global admin
        needs an explicit grant (for example Key Vault Purge Operator) in this
        vault. The default vault and purge-protected vaults are refused.
      operationId: purgeVault
      tags:
        - Vaults
      responses:
        "204":
          description: Vault purged successfully
        "400":
          description: The vault is default-protected or purge-protected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Role assignments ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/v1/users/login`,
        `POST /api/v1/users/refresh`, the OIDC callback, or
        `POST /api/v1/oauth2/token`. Tokens are signed asymmetrically
        (RS256/ES256) and always carry a `kid` header.
  parameters:
    VaultName:
      name: vault_name
      in: path
      required: true
      description: Name of the target vault
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
        example: "production"
...
  responses:
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or the session is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The caller is not authorized for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - id
        - message
        - status_code
      properties:
        id:
          type: string
          description: Error identifier
          example: "api.secret.create"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request body"
        detailed_error:
          type: string
          description: Detailed error information
          example: "json: invalid character '}' looking for beginning of value"
        status_code:
          type: integer
          description: HTTP status code
          example: 400
...
    CreateVaultRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: 3-63 lowercase alphanumerics or hyphens, no leading or trailing hyphen
          pattern: "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$"
          example: "production"
        enabled:
          type: boolean
        purge_protection:
          type: boolean
        retention_days:
          type: integer
        tags:
          type: object
          description: At most 15 entries; keys and values are 1-256 characters
          additionalProperties:
            type: string
    UpdateVaultRequest:
      type: object
      description: Omitted fields are left unchanged.
      properties:
        enabled:
          type: boolean
        purge_protection:
          type: boolean
        retention_days:
          type: integer
        tags:
          type: object
          description: Omitted leaves tags unchanged, `{}` clears them, a value replaces them
          additionalProperties:
            type: string
    VaultResponse:
      type: object
      required:
        - id
        - name
        - enabled
        - purge_protection
        - retention_days
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        purge_protection:
          type: boolean
        retention_days:
          type: integer
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
        scheduled_purge_at:
          type: string
          format: date-time
        tags:
          type: object
          additionalProperties:
            type: string
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          format: uuid
    ListVaultsResponse:
      type: object
      required:
        - vaults
        - total
      properties:
        vaults:
          type: array
          items:
            $ref: "#/components/schemas/VaultResponse"
        total:
          type: integer
    UpsertVaultWebhookRequest:
      type: object
      required:
        - url
      description: |
        There is no client-supplied secret field — the server always mints
        the signing secret.
      properties:
        url:
          type: string
          format: uri
          description: Must be an absolute https URL with a host.
          example: "https://hooks.example.com/rocketvault"
        rotate_secret:
          type: boolean
          description: Replace the existing signing secret with a newly minted one.
          default: false
        enabled:
          type: boolean
          description: Omitted leaves the current value unchanged (defaults to `true` on create).
    VaultWebhookConfigResponse:
      type: object
      description: |
        Has no property for the signing secret in any form, encrypted or
        plain — it is never returned by this shape.
      required:
        - url
        - enabled
        - created_at
        - updated_at
      properties:
        url:
          type: string
          format: uri
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VaultWebhookConfigCreatedResponse:
      type: object
      description: |
        VaultWebhookConfigResponse plus the plaintext signing secret, shown
        exactly once: only in the PUT response for the call that minted it
        (create, or rotate). No other endpoint or later call ever returns
        this field again.
      allOf:
        - $ref: "#/components/schemas/VaultWebhookConfigResponse"
        - type: object
          required:
            - signing_secret
          properties:
            signing_secret:
              type: string
              description: Plaintext signing secret. Shown once; not retrievable afterwards.

    # ─── Role assignments ────────────────────────────────────────────────────
...